Refractor with database changes
